Finalized benchmarks
Sample data generated at instanciation time of the benchmark class because the library was giving me headaches to either lazy constructor random sample data or directly set every indices of the collections (list/array).
Change the size of the sample data to get to the point quicker
Generation of data sample data is now completely removed from all the benchmarks except for when a collection has to be generated on the fly and we want to measure the time it takes to generate a sequence with unique values.