Fix Dotnet Pack
This gets rid of the need for .nuspec
and uses the csproj file. It also makes
sure that all resources for the analyzer are
in the nuget package.

As a bonus I upped the Test project to netcoreapp3.1 and
added a global.json file to keep things in the 3 range
for folks with multiple SDKs on their machines.